+++ src/navigation/AppNavigator.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+import { NavigationContainer } from '@react-navigation/native';
+/* screen */
+import { HomeScreen } from '../screens/HomeScreen'
+
+export const AppNavigator = () => {
+    return (
+        <NavigationContainer>
+            <HomeScreen />
+        </NavigationContainer>
+    )
+}
